found the problem (double free related) and resolved it
look at the changes made in this commit. addBook(Book b) or mergeShop(BookShop b). As Khaled Sir pointed out in the class, copy constructor is no more called now when an object is passed as an argument to a function or returned from a function. It is only called in cases of initialization. Because of copy constructor not getting called, the char *title is attempted to free more than once. Other clues which led to this discovery of mine was that things were perfectly ok with char title[100] but not with char *title. Will google for the exact documentation of modern compilers as to why copy constructor is no more called in these cases, might add the link to this comment if I can.

All in all, passing reference solves all the problems quite nicely.